Extract sendToWorker function
